test(spec): extract lifecycle drift guards to dedicated test file
Refs #1926. Test file size grew to 466 lines after batch 2 fixes
(approaching 500-line taste-lint warning). Drift-guard tests verify
config files (.markdownlint-cli2.yaml, .githooks/pre-commit), not
spec.md content, so they belong in a dedicated module.

- New: tests/commands/test_lifecycle_command_drift.py (69 lines, 3 tests)
- test_spec_step0.py drops to 403 lines (still warning, but spec-only)
- 55 tests still passing
