+++ tests/commands/test_spec_step0.py
@@ -118,52 +118,6 @@ def test_auto_mode_halt_token_in_skill_md(skill_text: str) -> None:
     assert "STEP_0_REQUIRES_ELICITATION" in skill_text
 
 
-# ---------------------------------------------------------------------------
-# Drift guard: lifecycle command paths in `.claude/commands/` must match
-# the markdownlint and skill-validator exclusions. Adding a 7th lifecycle
-# command without updating the two exclusion lists silently re-includes
-# the new file in CI lint and validator passes — those passes will fail
-# on the file's lifecycle-shape frontmatter. This test catches the drift.
-# ---------------------------------------------------------------------------
-
-
-COMMANDS_DIR = PROJECT_ROOT / ".claude" / "commands"
-MARKDOWNLINT_CONFIG = PROJECT_ROOT / ".markdownlint-cli2.yaml"
-PRE_COMMIT_HOOK = PROJECT_ROOT / ".githooks" / "pre-commit"
-
-# These lifecycle commands have YAML frontmatter + `@CLAUDE.md` body shape
-# (no H1, no Triggers/Verification sections). They are excluded from
-# markdownlint MD041 and SkillForge structural validation. Other command
-# files (.claude/commands/*.md) follow normal markdown conventions.
-LIFECYCLE_COMMANDS = {"spec", "plan", "build", "test", "review", "ship"}
-
-
-def test_lifecycle_commands_exist_in_claude_commands_dir() -> None:
-    """Every lifecycle command name must have a corresponding .md file."""
-    for cmd in LIFECYCLE_COMMANDS:
-        path = COMMANDS_DIR / f"{cmd}.md"
-        assert path.exists(), f"missing {path}"
-
-
-def test_markdownlint_excludes_match_lifecycle_commands() -> None:
-    """`.markdownlint-cli2.yaml` ignores list must include every lifecycle command twice:
-    once for `.claude/commands/<name>.md` and once for the Copilot CLI
-    mirror at `src/copilot-cli/skills/<name>/SKILL.md`. A 7th lifecycle
-    command added to `.claude/commands/` without two corresponding
-    ignores entries is drift.
-    """
-    config = MARKDOWNLINT_CONFIG.read_text(encoding="utf-8")
-    for cmd in LIFECYCLE_COMMANDS:
-        claude_path = f'.claude/commands/{cmd}.md"'
-        copilot_path = f'src/copilot-cli/skills/{cmd}/SKILL.md"'
-        assert claude_path in config, (
-            f"markdownlint ignores missing entry for {cmd} (Claude Code): {claude_path}"
-        )
-        assert copilot_path in config, (
-            f"markdownlint ignores missing entry for {cmd} (Copilot CLI mirror): {copilot_path}"
-        )
-
-
 def test_halt_emission_format_present(spec_text: str) -> None:
     """REQ-006-12 + Gate 5 #2: spec.md MUST define the machine-readable
     `step0-halt` fenced-block format (info-string + 5 keys). Free-form
@@ -201,23 +155,6 @@ def test_q1_q5_differentiation_in_spec_md(spec_text: str) -> None:
     )
 
 
-def test_pre_commit_hook_excludes_match_lifecycle_commands() -> None:
-    """`.githooks/pre-commit` skill-validator filter must include every
-    lifecycle command in the Copilot CLI exclusion regex. A 7th lifecycle
-    command not in the regex is silently re-included in skill validation.
-    """
-    hook = PRE_COMMIT_HOOK.read_text(encoding="utf-8")
-    # The regex is on a known line; extract the alternation group.
-    match = re.search(
-        r"src/copilot-cli/skills/\(([\w|]+)\)/SKILL\\\.md",
-        hook,
-    )
-    assert match is not None, "pre-commit hook lifecycle exclusion regex not found"
-    regex_commands = set(match.group(1).split("|"))
-    assert regex_commands == LIFECYCLE_COMMANDS, (
-        f"pre-commit hook exclusion regex {regex_commands} != "
-        f"canonical lifecycle commands {LIFECYCLE_COMMANDS}"
-    )
 
 
 # ---------------------------------------------------------------------------
